+++ tipg/factory.py
@@ -232,6 +232,7 @@ class OGCFeaturesFactory(EndpointsFactory):
     """OGC Features Endpoints Factory."""
 
     full: bool = False
+    tags: Optional[List[str]] = None
 
     @property
     def conforms_to(self) -> List[str]:
@@ -261,6 +262,7 @@ class OGCFeaturesFactory(EndpointsFactory):
                     }
                 },
             },
+            tags=self.tags,
         )
         def collections(
             request: Request,
@@ -426,6 +428,7 @@ class OGCFeaturesFactory(EndpointsFactory):
                     }
                 },
             },
+            tags=self.tags,
         )
         def collection(
             request: Request,
@@ -512,6 +515,7 @@ class OGCFeaturesFactory(EndpointsFactory):
                     }
                 },
             },
+            tags=self.tags,
         )
         def queryables(
             request: Request,
@@ -560,6 +564,7 @@ class OGCFeaturesFactory(EndpointsFactory):
                     "model": model.Items,
                 },
             },
+            tags=self.tags,
         )
         async def items(
             request: Request,
@@ -817,6 +822,7 @@ class OGCFeaturesFactory(EndpointsFactory):
                     "model": model.Item,
                 },
             },
+            tags=self.tags,
         )
         async def item(
             request: Request,
@@ -1121,6 +1127,7 @@ class OGCTilesFactory(EndpointsFactory):
 
     # OGC Tiles dependency
     supported_tms: TileMatrixSets = default_tms
+    tags: Optional[List[str]] = None
 
     @property
     def conforms_to(self) -> List[str]:
@@ -1138,11 +1145,13 @@ class OGCTilesFactory(EndpointsFactory):
             "/collections/{collectionId}/tiles/{tileMatrixSetId}/{tileMatrix}/{tileCol}/{tileRow}",
             response_class=Response,
             responses={200: {"content": {MediaType.mvt.value: {}}}},
+            tags=self.tags,
         )
         @self.router.get(
             "/collections/{collectionId}/tiles/{tileMatrix}/{tileCol}/{tileRow}",
             response_class=Response,
             responses={200: {"content": {MediaType.mvt.value: {}}}},
+            tags=self.tags,
         )
         async def tile(
             request: Request,
@@ -1202,12 +1211,14 @@ class OGCTilesFactory(EndpointsFactory):
             response_model=model.TileJSON,
             responses={200: {"description": "Return a tilejson"}},
             response_model_exclude_none=True,
+            tags=self.tags,
         )
         @self.router.get(
             "/collections/{collectionId}/tilejson.json",
             response_model=model.TileJSON,
             responses={200: {"description": "Return a tilejson"}},
             response_model_exclude_none=True,
+            tags=self.tags,
         )
         async def tilejson(
             request: Request,
@@ -1281,9 +1292,12 @@ class OGCTilesFactory(EndpointsFactory):
         @self.router.get(
             "/collections/{collectionId}/{tileMatrixSetId}/viewer",
             response_class=HTMLResponse,
+            tags=self.tags,
         )
         @self.router.get(
-            "/collections/{collectionId}/viewer", response_class=HTMLResponse
+            "/collections/{collectionId}/viewer",
+            response_class=HTMLResponse,
+            tags=self.tags,
         )
         def viewer_endpoint(
             request: Request,
@@ -1331,6 +1345,7 @@ class OGCTilesFactory(EndpointsFactory):
             response_model_exclude_none=True,
             summary="Retrieve the list of available tiling schemes (tile matrix sets).",
             operation_id="getTileMatrixSetsList",
+            tags=self.tags,
         )
         async def tilematrixsets(request: Request):
             """
@@ -1363,6 +1378,7 @@ class OGCTilesFactory(EndpointsFactory):
             response_model_exclude_none=True,
             summary="Retrieve the definition of the specified tiling scheme (tile matrix set).",
             operation_id="getTileMatrixSet",
+            tags=self.tags,
         )
         async def tilematrixset(
             tileMatrixSetId: Literal[tuple(self.supported_tms.list())] = Path(
@@ -1374,3 +1390,220 @@ class OGCTilesFactory(EndpointsFactory):
             OGC Specification: http://docs.opengeospatial.org/per/19-069.html#_tilematrixset
             """
             return self.supported_tms.get(tileMatrixSetId)
+
+
+@dataclass
+class Endpoints(EndpointsFactory):
+    """OGC Features and Tiles Endpoints Factory."""
+
+    name: str = "TiPg: OGC Features and Tiles API"
+
+    # OGC Tiles dependency
+    supported_tms: TileMatrixSets = default_tms
+
+    ogc_features: OGCFeaturesFactory = field(init=False)
+    ogc_tiles: OGCTilesFactory = field(init=False)
+
+    def __post_init__(self):
+        """Post Init: register route and configure specific options."""
+        self.ogc_features = OGCFeaturesFactory(
+            router=self.router,
+            collection_dependency=self.collection_dependency,
+            router_prefix=self.router_prefix,
+            templates=self.templates,
+            full=False,
+            tags=["OGC Features API"],
+        )
+        self.ogc_tiles = OGCTilesFactory(
+            router=self.router,
+            collection_dependency=self.collection_dependency,
+            router_prefix=self.router_prefix,
+            templates=self.templates,
+            supported_tms=self.supported_tms,
+            tags=["OGC Tiles API"],
+        )
+        self.register_routes()
+
+    @property
+    def conforms_to(self) -> List[str]:
+        """Endpoints conformances."""
+        return [
+            # OGC Common
+            "http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/core",
+            "http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/landingPage",
+            "http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/collections",
+            "http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/simple-query",
+            "http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/json",
+            "http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/html",
+            "http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/oas30",
+            *self.ogc_features.conforms_to,
+            *self.ogc_tiles.conforms_to,
+        ]
+
+    def register_routes(self):
+        """Register factory Routes."""
+
+        @self.router.get(
+            "/conformance",
+            response_model=model.Conformance,
+            response_model_exclude_none=True,
+            response_class=ORJSONResponse,
+            responses={
+                200: {
+                    "content": {
+                        MediaType.json.value: {},
+                        MediaType.html.value: {},
+                    }
+                },
+            },
+        )
+        def conformance(
+            request: Request, output_type: Optional[MediaType] = Depends(OutputType)
+        ):
+            """Get conformance."""
+            data = model.Conformance(conformsTo=self.conforms_to)
+
+            if output_type == MediaType.html:
+                return create_html_response(
+                    request,
+                    data.json(exclude_none=True),
+                    templates=self.templates,
+                    template_name="conformance",
+                )
+
+            return data
+
+        @self.router.get(
+            "/",
+            response_model=model.Landing,
+            response_model_exclude_none=True,
+            response_class=ORJSONResponse,
+            responses={
+                200: {
+                    "content": {
+                        MediaType.json.value: {},
+                        MediaType.html.value: {},
+                    }
+                },
+            },
+        )
+        def landing(
+            request: Request, output_type: Optional[MediaType] = Depends(OutputType)
+        ):
+            """Get landing page."""
+            data = model.Landing(
+                title=self.name,
+                links=[
+                    model.Link(
+                        title="Landing Page",
+                        href=str(request.url_for("landing")),
+                        type=MediaType.json,
+                        rel="self",
+                    ),
+                    model.Link(
+                        title="the API definition (JSON)",
+                        href=str(request.url_for("openapi")),
+                        type=MediaType.openapi30_json,
+                        rel="service-desc",
+                    ),
+                    model.Link(
+                        title="the API documentation",
+                        href=str(request.url_for("swagger_ui_html")),
+                        type=MediaType.html,
+                        rel="service-doc",
+                    ),
+                    model.Link(
+                        title="Conformance",
+                        href=self.url_for(request, "conformance"),
+                        type=MediaType.json,
+                        rel="conformance",
+                    ),
+                    model.Link(
+                        title="List of Collections",
+                        href=self.url_for(request, "collections"),
+                        type=MediaType.json,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="Collection metadata",
+                        href=self.url_for(
+                            request,
+                            "collection",
+                            collectionId="{collectionId}",
+                        ),
+                        type=MediaType.json,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="Collection queryables",
+                        href=self.url_for(
+                            request,
+                            "queryables",
+                            collectionId="{collectionId}",
+                        ),
+                        type=MediaType.schemajson,
+                        rel="queryables",
+                    ),
+                    model.Link(
+                        title="Collection Features",
+                        href=self.url_for(
+                            request, "items", collectionId="{collectionId}"
+                        ),
+                        type=MediaType.geojson,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="Collection Vector Tiles",
+                        href=self.url_for(
+                            request,
+                            "tile",
+                            collectionId="{collectionId}",
+                            tileMatrix="{tileMatrix}",
+                            tileCol="{tileCol}",
+                            tileRow="{tileRow}",
+                        ),
+                        type=MediaType.mvt,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="Collection Feature",
+                        href=self.url_for(
+                            request,
+                            "item",
+                            collectionId="{collectionId}",
+                            itemId="{itemId}",
+                        ),
+                        type=MediaType.geojson,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="TileMatrixSets",
+                        href=self.url_for(
+                            request,
+                            "tilematrixsets",
+                        ),
+                        type=MediaType.json,
+                        rel="data",
+                    ),
+                    model.Link(
+                        title="TileMatrixSet",
+                        href=self.url_for(
+                            request,
+                            "tilematrixset",
+                            tileMatrixSetId="{tileMatrixSetId}",
+                        ),
+                        type=MediaType.json,
+                        rel="data",
+                    ),
+                ],
+            )
+
+            if output_type == MediaType.html:
+                return create_html_response(
+                    request,
+                    data.json(exclude_none=True),
+                    templates=self.templates,
+                    template_name="landing",
+                )
+
+            return data
